requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
black==22.10.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.3
click==8.1.3
cloudinary==1.26.0
colorama==0.4.6
config==0.5.0.post0
confusable-homoglyphs==3.2.0
coverage==5.3
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.16
django-appconf==1.0.4
django-bootstrap-v5==1.0.11
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-decouple==2.1
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-imagekit==4.0.2
django-jquery==3.1.0
django-material==1.9.0
django-phonenumber-field==7.0.2
django-registration==3.2
django-rest-knox==4.1.0
django-tinymce==3.5.0
django-url-or-relative-url-field==0.1.2
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.2
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.13.6
pilkit==2.0
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.20
PyJWT>=1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==1.82.1
sweetify==2.3.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.3
waitress==2.1.2
whitenoise==5.2.0
